+++ MinimedKit/PumpEvents/CalBGForPHPumpEvent.swift
@@ -23,8 +23,8 @@ public struct CalBGForPHPumpEvent: TimestampedPumpEvent {
 
         rawData = availableData.subdata(in: 0..<length)
         
-        func d(_ idx:Int) -> Int {
-            return Int(availableData[idx] as UInt8)
+        func d(_ idx: Int) -> Int {
+            return Int(availableData[idx])
         }
         
         timestamp = DateComponents(pumpEventData: availableData, offset: 2)
